+++ api/src/api/services/rates.service.ts
@@ -6,9 +6,7 @@ import { __dirname } from "../../config";
 export const handleGetRates = async (ctx: Context) => {
   const tempDir = path.join(__dirname, `/../temp`);
 
-  const resultJson = await Bun.file(
-    path.join(tempDir, `/exchange_rates.json`)
-  ).json();
+  const resultJson = await Bun.file(path.join(tempDir, `/exchange_rates.json`)).json();
 
   if (!resultJson) {
     ctx.set.status = 400;