+++ web/src/modules/Search/components/TabVideos/index.tsx
@@ -1,6 +1,6 @@
 import { Button, Center, Flex, SimpleGrid, Text } from "@mantine/core";
 import type { ISearXNGResultsVideos } from "@ts/searxng.types";
-import { useEffect } from "react";
+import { useEffect, useState } from "react";
 import useSearXNGSWR from "src/api/searxng/use-searxng-query";
 import VideoCell from "./components/VideoCell";
 import VideoSkeleton from "./components/VideoSkeleton";
@@ -12,6 +12,8 @@ import PrivatePlayer from "../components/PrivatePlayer";
 
 const TabVideos = () => {
   const hydrated = useEnginesStore((state) => state.hydrated);
+
+  const [privatePlayerURL, setPrivatePlayerURL] = useState("");
 
   const { data, error, isLoading, isValidating, setSize, size, mutate } =
     useSearXNGSWR<ISearXNGResultsVideos>();
@@ -39,7 +41,9 @@ const TabVideos = () => {
       >
         {data?.map((res) => {
           if (!res) return;
-          return res?.results.map((img, i) => <VideoCell key={i} videoData={img} />);
+          return res?.results.map((vid, i) => (
+            <VideoCell key={i} videoData={vid} setPrivatePlayerURL={setPrivatePlayerURL} />
+          ));
         })}
 
         {(isLoading || isValidating || !hydrated) &&
@@ -63,7 +67,12 @@ const TabVideos = () => {
       )}
 
       {/* Private Player */}
-      {/* <PrivatePlayer /> */}
+      <PrivatePlayer
+        url={privatePlayerURL}
+        onClose={() => {
+          setPrivatePlayerURL("");
+        }}
+      />
     </Flex>
   );
 };