+++ web/src/modules/Search/components/components/Organize/components/OBlacklist.tsx
@@ -2,7 +2,7 @@ import classes from "../styles.module.scss";
 import { Accordion, Center, SimpleGrid, Text } from "@mantine/core";
 import { useSearchStore } from "@store/search";
 import { IconForbid } from "@tabler/icons-react";
-import ResultItem from "./ResultItem";
+import DomainItem from "./DomainItem";
 
 const OBlacklist = () => {
   const domainsBlacklist = useSearchStore((state) => state.domainsBlacklist);
@@ -16,7 +16,7 @@ const OBlacklist = () => {
         {domainsBlacklist.length ? (
           <SimpleGrid mt="lg" cols={1} spacing="md">
             {domainsBlacklist.map((item, i) => (
-              <ResultItem key={i} domain={item} isCurrent={false} />
+              <DomainItem key={i} domain={item} isCurrent={false} />
             ))}
           </SimpleGrid>
         ) : (