+++ web/src/modules/Settings/components/_instances/AIWorker/index.tsx
@@ -1,19 +1,21 @@
 import { Anchor, Button, Flex, Paper, Stack, Text, TextInput } from "@mantine/core";
-import { IconWorld } from "@tabler/icons-react";
+import { IconSparkles, IconWorld } from "@tabler/icons-react";
 
-import classes from "../../styles.module.scss";
+import classes from "../../../styles.module.scss";
 import { getIconStyle } from "@utils/functions/iconStyle";
 import useToast from "@hooks/use-toast";
 import { useInstanceStore } from "@store/instance";
 import useForm from "@hooks/use-form";
 import { useEffect } from "react";
-import ExternalLink from "@components/ExternalLink";
+import RemixLink from "@components/RemixLink";
 import { usePrimaryColor } from "@hooks/use-primary-color";
 
-const SettingsNominatim = () => {
+const SettingsAIWorker = () => {
   const hydrated = useInstanceStore((state) => state.hydrated);
-  const domain = useInstanceStore((state) => state.nominatimDomain);
-  const setDomain = useInstanceStore((state) => state.setNominatimDomain);
+  const domain = useInstanceStore((state) => state.workerDomain);
+  const setDomain = useInstanceStore((state) => state.setWorkerDomain);
+
+  const linkTextColor = usePrimaryColor(4);
 
   const form = useForm({
     initialValues: {
@@ -25,8 +27,6 @@ const SettingsNominatim = () => {
   });
 
   const { toast } = useToast();
-
-  const linkTextColor = usePrimaryColor(4);
 
   const handleSubmit = (values: typeof form.values) => {
     setDomain(values.domain);
@@ -43,10 +43,10 @@ const SettingsNominatim = () => {
     <Paper radius="md" mt={40} withBorder>
       <form onSubmit={form.onSubmit(handleSubmit)}>
         <Flex align="center" p="lg" mb={16}>
-          <IconWorld style={getIconStyle(32)} />
+          <IconSparkles style={getIconStyle(32)} />
 
           <Text fz={26} fw={600} ml="sm">
-            Nominatim Domain
+            AI Worker
           </Text>
         </Flex>
 
@@ -60,8 +60,17 @@ const SettingsNominatim = () => {
           />
 
           <Text size="sm">
-            <ExternalLink href="https://nominatim.org/">Nominatim</ExternalLink> uses OpenStreetMap
-            data to find locations on Earth by name and address (geocoding).
+            <Anchor
+              href="https://developers.cloudflare.com/workers-ai/get-started/dashboard/"
+              target="_blank"
+              rel="noreferrer noopener"
+            >
+              <Text component="span" c={linkTextColor}>
+                Cloudflare AI Workers
+              </Text>
+            </Anchor>{" "}
+            are used to display AI answers in /search. Click read more below to learn how to set up
+            your own worker.
           </Text>
         </Stack>
 
@@ -74,11 +83,11 @@ const SettingsNominatim = () => {
         >
           <Text size="sm" c="dimmed">
             Change this to your own url for better privacy & less load for default instance.{" "}
-            <Text component="span" c={linkTextColor}>
-              <Anchor href="https://nominatim.org/" target="_blank" rel="noreferrer noopener">
+            <RemixLink to={"/docs/self-host-ai-worker"}>
+              <Text component="span" c={linkTextColor}>
                 Read more
-              </Anchor>
-            </Text>
+              </Text>
+            </RemixLink>
           </Text>
 
           <Button type="submit">Save</Button>
@@ -88,4 +97,4 @@ const SettingsNominatim = () => {
   );
 };
 
-export default SettingsNominatim;
+export default SettingsAIWorker;