+++ web/src/modules/Settings/components/_instances/Nominatim/index.tsx
@@ -1,21 +1,19 @@
 import { Anchor, Button, Flex, Paper, Stack, Text, TextInput } from "@mantine/core";
-import { IconSparkles, IconWorld } from "@tabler/icons-react";
+import { IconWorld } from "@tabler/icons-react";
 
-import classes from "../../styles.module.scss";
+import classes from "../../../styles.module.scss";
 import { getIconStyle } from "@utils/functions/iconStyle";
 import useToast from "@hooks/use-toast";
 import { useInstanceStore } from "@store/instance";
 import useForm from "@hooks/use-form";
 import { useEffect } from "react";
-import RemixLink from "@components/RemixLink";
+import ExternalLink from "@components/ExternalLink";
 import { usePrimaryColor } from "@hooks/use-primary-color";
 
-const SettingsAIWorker = () => {
+const SettingsNominatim = () => {
   const hydrated = useInstanceStore((state) => state.hydrated);
-  const domain = useInstanceStore((state) => state.workerDomain);
-  const setDomain = useInstanceStore((state) => state.setWorkerDomain);
-
-  const linkTextColor = usePrimaryColor(4);
+  const domain = useInstanceStore((state) => state.nominatimDomain);
+  const setDomain = useInstanceStore((state) => state.setNominatimDomain);
 
   const form = useForm({
     initialValues: {
@@ -27,6 +25,8 @@ const SettingsAIWorker = () => {
   });
 
   const { toast } = useToast();
+
+  const linkTextColor = usePrimaryColor(4);
 
   const handleSubmit = (values: typeof form.values) => {
     setDomain(values.domain);
@@ -43,10 +43,10 @@ const SettingsAIWorker = () => {
     <Paper radius="md" mt={40} withBorder>
       <form onSubmit={form.onSubmit(handleSubmit)}>
         <Flex align="center" p="lg" mb={16}>
-          <IconSparkles style={getIconStyle(32)} />
+          <IconWorld style={getIconStyle(32)} />
 
           <Text fz={26} fw={600} ml="sm">
-            AI Worker
+            Nominatim Domain
           </Text>
         </Flex>
 
@@ -60,17 +60,8 @@ const SettingsAIWorker = () => {
           />
 
           <Text size="sm">
-            <Anchor
-              href="https://developers.cloudflare.com/workers-ai/get-started/dashboard/"
-              target="_blank"
-              rel="noreferrer noopener"
-            >
-              <Text component="span" c={linkTextColor}>
-                Cloudflare AI Workers
-              </Text>
-            </Anchor>{" "}
-            are used to display AI answers in /search. Click read more below to learn how to set up
-            your own worker.
+            <ExternalLink href="https://nominatim.org/">Nominatim</ExternalLink> uses OpenStreetMap
+            data to find locations on Earth by name and address (geocoding).
           </Text>
         </Stack>
 
@@ -83,11 +74,11 @@ const SettingsAIWorker = () => {
         >
           <Text size="sm" c="dimmed">
             Change this to your own url for better privacy & less load for default instance.{" "}
-            <RemixLink to={"/docs/self-host-ai-worker"}>
-              <Text component="span" c={linkTextColor}>
+            <Text component="span" c={linkTextColor}>
+              <Anchor href="https://nominatim.org/" target="_blank" rel="noreferrer noopener">
                 Read more
-              </Text>
-            </RemixLink>
+              </Anchor>
+            </Text>
           </Text>
 
           <Button type="submit">Save</Button>
@@ -97,4 +88,4 @@ const SettingsAIWorker = () => {
   );
 };
 
-export default SettingsAIWorker;
+export default SettingsNominatim;