+++ CHANGELOG.md
@@ -25,7 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Instant answers - Time In, Time Zone
 - New and improved theme customization
 - Multiple missing search engines
-- Media Player
+- Private Media Player
 
 ### Fixed